// types/site-settings.ts
export type SiteSettings = {
  contactEmail: string
  contactPhone: string
  facebookUrl: string
  instagramUrl: string
  youtubeUrl: string
  heroVariantId: string
  maintenanceMode: boolean
  maintenanceMessageKa: string
  maintenanceMessageEn: string
  maintenanceMessageRu: string
}

export const SITE_SETTINGS_DOC_ID = 'main'